+++ package.json
@@ -16,7 +16,7 @@
     "@sanity/image-url": "^1.0.1",
     "bootstrap": "^5.1.3",
     "isomorphic-fetch": "^3.0.0",
-    "next": "^12.1.0",
+    "next": "^12.2.2",
     "next-sanity-image": "^3.2.1",
     "npm": "^8.5.2",
     "react": "17.0.2",
@@ -29,8 +29,8 @@
   "devDependencies": {
     "@kyleshevlin/eslint-plugin": "^1.3.0",
     "@types/isomorphic-fetch": "^0.0.35",
-    "@types/node": "17.0.14",
-    "@types/react": "17.0.38",
+    "@types/node": "^17.0.14",
+    "@types/react": "^17.0.38",
     "@typescript-eslint/eslint-plugin": "^5.9.1",
     "eslint": "8.8.0",
     "eslint-config-next": "12.0.10",
@@ -43,7 +43,7 @@
     "husky": "^7.0.4",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.5.1",
-    "typescript": "4.5.5"
+    "typescript": "^4.5.5"
   },
   "husky": {
     "hooks": {
