+++ styles/pages/HighPointPark.styles.module.scss
@@ -146,6 +146,13 @@
   font-weight: light;
   color: #0c256c;
 }
+
+// project video
+.projectRowVideo {
+  padding-top: 101px;
+}
+
+// project row two
 .projectRowTwo {
   padding-top: 101px;
 }
